Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["limahgustavo <gustavo.limafgz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
pymysql = "^1.1.1"
cryptography = "^42.0.8"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-decouple = "^3.8"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
